separe some functions
